requirements.txt
flask
numpy
soundfile
python-dotenv
# Optional, not currently used but available
webrtcvad
vosk
